refactor(api): decompose rest_api.py into handler mixins and data models
Extracts 7 handler mixin classes (core, generation, conversion, validation,
inertia, library, editor), a models module (HTTPMethod, APIRequest, APIResponse,
Route), and a framework adapters module (FlaskAdapter, FastAPIAdapter).
rest_api.py shrinks from 1290 → 398 lines; all public imports remain stable
via re-exports. Removes the rest_api.py exception from file_size_budget.json.
Adds 43 new TDD tests (209 total, all passing).
