+++ scripts/config/file_size_budget.json
@@ -62,12 +62,6 @@
       "expires_on": "2026-12-31"
     },
     {
-      "path": "src/shared/python/model_generation/api/rest_api.py",
-      "owner": "@model-generation",
-      "reason": "REST API grew past 1200 lines due to security features (auth, CORS, rate limiting). Decomposition into handler modules planned.",
-      "expires_on": "2027-03-09"
-    },
-    {
       "path": "src/shared/python/humanoid_character_builder/generators/mesh_generator.py",
       "owner": "@humanoid-builder",
       "reason": "Pre-existing over-budget file (1641 lines on main). Geometry generation split pending.",
