+++ src/shared/python/model_generation/api/handlers/__init__.py
@@ -1,0 +1,19 @@
+"""Handler mixin modules for ModelGenerationAPI."""
+
+from model_generation.api.handlers.conversion import ConversionHandlersMixin
+from model_generation.api.handlers.core import CoreHandlersMixin
+from model_generation.api.handlers.editor import EditorHandlersMixin
+from model_generation.api.handlers.generation import GenerationHandlersMixin
+from model_generation.api.handlers.inertia import InertiaHandlersMixin
+from model_generation.api.handlers.library import LibraryHandlersMixin
+from model_generation.api.handlers.validation import ValidationHandlersMixin
+
+__all__ = [
+    "CoreHandlersMixin",
+    "GenerationHandlersMixin",
+    "ConversionHandlersMixin",
+    "ValidationHandlersMixin",
+    "InertiaHandlersMixin",
+    "LibraryHandlersMixin",
+    "EditorHandlersMixin",
+]
